fix(aiguard): Remove aiguard product code from instrumentation
