+++ packages/datadog-instrumentations/src/openai.js
@@ -3,10 +3,51 @@
 const dc = require('dc-polyfill')
 const shimmer = require('../../datadog-shimmer')
 const { addHook } = require('./helpers/instrument')
-const aiGuard = require('./helpers/openai-ai-guard')
 
 const ch = dc.tracingChannel('apm:openai:request')
 const onStreamedChunkCh = dc.channel('apm:openai:request:chunk')
+
+// Provider lifecycle channels. Payloads stay OpenAI-native:
+// before { args, parentSpan, abortController, pending }
+// after  { args, body, parentSpan, abortController, pending }
+const chatCompletionsBeforeChannel = dc.channel('dd-trace:openai:chat.completions:before')
+const chatCompletionsAfterChannel = dc.channel('dd-trace:openai:chat.completions:after')
+const responsesBeforeChannel = dc.channel('dd-trace:openai:responses:before')
+const responsesAfterChannel = dc.channel('dd-trace:openai:responses:after')
+
+const LIFECYCLE_CHANNELS = {
+  'chat.completions': {
+    before: chatCompletionsBeforeChannel,
+    after: chatCompletionsAfterChannel,
+  },
+  responses: {
+    before: responsesBeforeChannel,
+    after: responsesAfterChannel,
+  },
+}
+
+/**
+ * Publishes a provider-native lifecycle payload to a cancelable lifecycle channel.
+ *
+ * Subscribers push async work into `pending` synchronously during publication and
+ * abort `abortController` with an error before the pushed promise resolves to block.
+ *
+ * @param {object} channel
+ * @param {object} payload
+ * @returns {Promise<void>}
+ */
+function publishLifecycle (channel, payload) {
+  const abortController = new AbortController()
+  const ctx = { ...payload, abortController, pending: [] }
+
+  channel.publish(ctx)
+
+  return Promise.all(ctx.pending).then(() => {
+    if (abortController.signal.aborted) {
+      throw abortController.signal.reason
+    }
+  })
+}
 
 const V4_PACKAGE_SHIMS = [
   {
@@ -217,17 +258,15 @@ for (const extension of extensions) {
 
       for (const methodName of methods) {
         shimmer.wrap(targetPrototype, methodName, methodFn => function (...args) {
-          if (!ch.start.hasSubscribers && !aiGuard.hasSubscribers()) {
-            return methodFn.apply(this, args)
-          }
           // The OpenAI library lets you set `stream: true` on the options arg to any method
           // However, we only want to handle streamed responses in specific cases
           // chat.completions and completions
           const stream = streamedResponse && getOption(args, 'stream', false)
 
-          const guard = aiGuard.createGuard(baseResource, args[0], stream)
+          const channels = stream ? null : LIFECYCLE_CHANNELS[baseResource]
+          const hasLifecycle = !!channels && (channels.before.hasSubscribers || channels.after.hasSubscribers)
 
-          if (!ch.start.hasSubscribers && !guard) {
+          if (!ch.start.hasSubscribers && !hasLifecycle) {
             return methodFn.apply(this, args)
           }
 
@@ -240,11 +279,18 @@ for (const extension of extensions) {
           }
 
           return ch.start.runStores(ctx, () => {
-            // Explicit childOf rather than async-context: the _thenUnwrap/parse path
-            // decouples the lazy evaluation from the active scope at call time.
-            if (guard) guard.parentSpan = ctx.currentStore?.span
+            // Capture the parent span explicitly: the _thenUnwrap/parse path decouples
+            // the lazy evaluation from the active scope at call time.
+            const parentSpan = hasLifecycle ? ctx.currentStore?.span : undefined
 
             const apiProm = methodFn.apply(this, args)
+
+            // Publish :before eagerly so the after gate and any side channels share one verdict.
+            const beforeVerdict = hasLifecycle && channels.before.hasSubscribers
+              ? publishLifecycle(channels.before, { args, parentSpan })
+              : null
+
+            const afterChannel = hasLifecycle && channels.after.hasSubscribers ? channels.after : null
 
             if (baseResource === 'chat.completions' && typeof apiProm._thenUnwrap === 'function') {
               // this should only ever be invoked from a client.beta.chat.completions.parse call
@@ -259,7 +305,7 @@ for (const extension of extensions) {
                   const parsedPromise = origApiPromParse.apply(this, args)
                     .then(body => Promise.all([this.responsePromise, body]))
 
-                  return handleUnwrappedAPIPromise(parsedPromise, ctx, stream, guard)
+                  return handleUnwrappedAPIPromise(parsedPromise, ctx, stream, beforeVerdict, afterChannel, parentSpan)
                 })
 
                 return unwrappedPromise
@@ -272,10 +318,16 @@ for (const extension of extensions) {
               const parsedPromise = origApiPromParse.apply(this, args)
                 .then(body => Promise.all([this.responsePromise, body]))
 
-              return handleUnwrappedAPIPromise(parsedPromise, ctx, stream, guard)
+              return handleUnwrappedAPIPromise(parsedPromise, ctx, stream, beforeVerdict, afterChannel, parentSpan)
             })
 
-            if (guard) aiGuard.wrapAsResponse(apiProm, guard)
+            // Gate `.asResponse()` callers on the before verdict so raw-response paths still block.
+            if (beforeVerdict && typeof apiProm.asResponse === 'function') {
+              shimmer.wrap(apiProm, 'asResponse', origAsResponse => function (...args) {
+                const responsePromise = origAsResponse.apply(this, args)
+                return Promise.all([beforeVerdict, responsePromise]).then(([, response]) => response)
+              })
+            }
 
             ch.end.publish(ctx)
 
@@ -288,10 +340,12 @@ for (const extension of extensions) {
   }
 }
 
-function handleUnwrappedAPIPromise (apiProm, ctx, stream, guard) {
-  const guardedApiProm = guard ? aiGuard.gateParse(apiProm, guard) : apiProm
+function handleUnwrappedAPIPromise (apiProm, ctx, stream, beforeVerdict, afterChannel, parentSpan) {
+  const gatedApiProm = beforeVerdict
+    ? Promise.all([beforeVerdict, apiProm]).then(([, result]) => result)
+    : apiProm
 
-  return guardedApiProm
+  return gatedApiProm
     .then(([{ response, options }, body]) => {
       if (stream) {
         if (body.iterator) {
@@ -313,14 +367,14 @@ function handleUnwrappedAPIPromise (apiProm, ctx, stream, guard) {
         },
       }
 
-      if (!guard) {
+      if (!afterChannel) {
         finish(ctx, responseData)
         return body
       }
 
       // Finish after evaluation so a block propagates the error to openai.request
-      // and the span wraps its ai_guard child instead of closing before it.
-      return aiGuard.evaluateOutput(guard, body).then(() => {
+      // and the span wraps its child instead of closing before it.
+      return publishLifecycle(afterChannel, { args: ctx.args, body, parentSpan }).then(() => {
         finish(ctx, responseData)
         return body
       })
