+++ packages/dd-trace/src/aiguard/channels.js
@@ -3,6 +3,5 @@
 const dc = require('dc-polyfill')
 
 module.exports = {
-  aiguardChannel: dc.channel('dd-trace:ai:aiguard'),
   incomingHttpRequestStart: dc.channel('dd-trace:incomingHttpRequestStart'),
 }
