+++ packages/dd-trace/src/aiguard/integrations/evaluate.js
@@ -1,0 +1,47 @@
+'use strict'
+
+const log = require('../../log')
+
+/**
+ * Pushes one AI Guard evaluation promise into a lifecycle ctx.
+ *
+ * Subscribers must push synchronously during channel publication. Abort before
+ * the pushed promise resolves so publishers can inspect `signal.reason` after
+ * `Promise.all(ctx.pending)`.
+ *
+ * @param {object} ctx
+ * @param {AbortController} ctx.abortController
+ * @param {object} [ctx.parentSpan]
+ * @param {Array<Promise<void>>} ctx.pending
+ * @param {object} aiguard
+ * @param {Array<object>|undefined} messages
+ * @param {object} opts
+ */
+function pushEvaluation (ctx, aiguard, messages, opts) {
+  if (!messages?.length) return
+
+  const evaluateOpts = ctx.parentSpan ? { ...opts, childOf: ctx.parentSpan } : opts
+
+  try {
+    ctx.pending.push(aiguard.evaluate(messages, evaluateOpts).catch(handleEvaluationError.bind(null, ctx)))
+  } catch (err) {
+    ctx.pending.push(Promise.resolve().then(() => handleEvaluationError(ctx, err)))
+  }
+}
+
+/**
+ * Handles an AI Guard evaluation failure.
+ *
+ * @param {object} ctx
+ * @param {AbortController} ctx.abortController
+ * @param {Error} err
+ */
+function handleEvaluationError (ctx, err) {
+  if (err.name === 'AIGuardAbortError') {
+    ctx.abortController.abort(err)
+  } else {
+    log.error('AIGuard: unexpected error during evaluation: %s', err.message)
+  }
+}
+
+module.exports = { pushEvaluation }
