+++ packages/dd-trace/test/aiguard/messages/openai.spec.js
@@ -1,0 +1,418 @@
+'use strict'
+
+const assert = require('node:assert/strict')
+const { describe, it } = require('mocha')
+
+const {
+  normalizeOpenAIChatMessages,
+  convertOpenAIResponseItemsToMessages,
+  convertOpenAIResponsePromptToMessages,
+  getResponsesInputMessages,
+  openAIResponseContentToMessageContent,
+} = require('../../../src/aiguard/messages/openai')
+
+describe('aiguard/messages/openai', () => {
+  describe('normalizeOpenAIChatMessages', () => {
+    it('should return undefined for unsupported or empty input', () => {
+      assert.strictEqual(normalizeOpenAIChatMessages(undefined), undefined)
+      assert.strictEqual(normalizeOpenAIChatMessages([]), undefined)
+    })
+
+    it('should preserve modern chat messages', () => {
+      const messages = [{
+        role: 'assistant',
+        tool_calls: [{ id: 'call_1', function: { name: 'lookup', arguments: '{}' } }],
+      }]
+
+      assert.deepStrictEqual(normalizeOpenAIChatMessages(messages), messages)
+    })
+
+    it('should convert deprecated assistant function_call messages to tool_calls', () => {
+      const messages = [{
+        role: 'assistant',
+        content: null,
+        function_call: { name: 'lookup', arguments: { query: 'test' } },
+      }]
+
+      assert.deepStrictEqual(normalizeOpenAIChatMessages(messages), [{
+        role: 'assistant',
+        content: null,
+        tool_calls: [{
+          id: 'lookup',
+          function: { name: 'lookup', arguments: '{"query":"test"}' },
+        }],
+      }])
+    })
+
+    it('should convert deprecated function role messages to tool messages', () => {
+      const messages = [{ role: 'function', name: 'lookup', content: { result: 'ok' } }]
+
+      assert.deepStrictEqual(normalizeOpenAIChatMessages(messages), [{
+        role: 'tool',
+        tool_call_id: 'lookup',
+        content: '{"result":"ok"}',
+      }])
+    })
+  })
+
+  describe('convertOpenAIResponseItemsToMessages', () => {
+    it('should convert string input to a default role message', () => {
+      assert.deepStrictEqual(convertOpenAIResponseItemsToMessages('Hello', 'user'), [
+        { role: 'user', content: 'Hello' },
+      ])
+    })
+
+    it('should return empty array for unsupported input', () => {
+      assert.deepStrictEqual(convertOpenAIResponseItemsToMessages(undefined, 'user'), [])
+      assert.deepStrictEqual(convertOpenAIResponseItemsToMessages({ input: 'Hello' }, 'user'), [])
+    })
+
+    it('should convert response message items to OpenAI chat-style messages', () => {
+      const items = [{
+        type: 'message',
+        role: 'user',
+        content: [{ type: 'input_text', text: 'Hello' }],
+      }]
+
+      assert.deepStrictEqual(convertOpenAIResponseItemsToMessages(items, 'user'), [
+        { role: 'user', content: 'Hello' },
+      ])
+    })
+
+    it('should use the default role when response message item has no role', () => {
+      const items = [{
+        type: 'message',
+        content: [{ type: 'output_text', text: 'Hi' }],
+      }]
+
+      assert.deepStrictEqual(convertOpenAIResponseItemsToMessages(items, 'assistant'), [
+        { role: 'assistant', content: 'Hi' },
+      ])
+    })
+
+    it('should preserve image URL content parts', () => {
+      const items = [{
+        type: 'message',
+        role: 'user',
+        content: [
+          { type: 'input_text', text: 'Describe this' },
+          { type: 'input_image', image_url: 'https://example.com/image.png' },
+        ],
+      }]
+
+      assert.deepStrictEqual(convertOpenAIResponseItemsToMessages(items, 'user'), [{
+        role: 'user',
+        content: [
+          { type: 'text', text: 'Describe this' },
+          { type: 'image_url', image_url: { url: 'https://example.com/image.png' } },
+        ],
+      }])
+    })
+
+    it('should convert function call items to assistant tool call messages', () => {
+      const items = [{
+        type: 'function_call',
+        call_id: 'call_1',
+        name: 'lookup',
+        arguments: { query: 'test' },
+      }]
+
+      assert.deepStrictEqual(convertOpenAIResponseItemsToMessages(items, 'assistant'), [{
+        role: 'assistant',
+        tool_calls: [{
+          id: 'call_1',
+          function: {
+            name: 'lookup',
+            arguments: '{"query":"test"}',
+          },
+        }],
+      }])
+    })
+
+    it('should convert function call output items to tool messages', () => {
+      const items = [{
+        type: 'function_call_output',
+        call_id: 'call_1',
+        output: { result: 'ok' },
+      }]
+
+      assert.deepStrictEqual(convertOpenAIResponseItemsToMessages(items, 'tool'), [{
+        role: 'tool',
+        tool_call_id: 'call_1',
+        content: '{"result":"ok"}',
+      }])
+    })
+
+    it('should preserve input_file content with a stable file marker or reference', () => {
+      const items = [{
+        type: 'message',
+        role: 'user',
+        content: [
+          { type: 'input_text', text: 'Read this' },
+          { type: 'input_file', file_id: 'file_123' },
+        ],
+      }]
+
+      assert.deepStrictEqual(convertOpenAIResponseItemsToMessages(items, 'user'), [{
+        role: 'user',
+        content: 'Read this\nfile_123',
+      }])
+    })
+
+    it('should convert custom tool call items to assistant tool call messages', () => {
+      const items = [{
+        type: 'custom_tool_call',
+        call_id: 'call_custom',
+        name: 'python',
+        input: 'print(1)',
+      }]
+
+      assert.deepStrictEqual(convertOpenAIResponseItemsToMessages(items, 'assistant'), [{
+        role: 'assistant',
+        tool_calls: [{
+          id: 'call_custom',
+          function: { name: 'python', arguments: 'print(1)' },
+        }],
+      }])
+    })
+
+    it('should convert custom tool call output items to tool messages', () => {
+      const items = [{
+        type: 'custom_tool_call_output',
+        call_id: 'call_custom',
+        output: [{ type: 'input_text', text: 'done' }],
+      }]
+
+      assert.deepStrictEqual(convertOpenAIResponseItemsToMessages(items, 'tool'), [{
+        role: 'tool',
+        tool_call_id: 'call_custom',
+        content: 'done',
+      }])
+    })
+
+    it('should convert MCP call items with output to linked tool call and tool output messages', () => {
+      const items = [{
+        type: 'mcp_call',
+        id: 'mcp_1',
+        name: 'search_docs',
+        server_label: 'docs',
+        arguments: '{"q":"x"}',
+        output: 'found it',
+      }]
+
+      assert.deepStrictEqual(convertOpenAIResponseItemsToMessages(items, 'assistant'), [
+        {
+          role: 'assistant',
+          tool_calls: [{
+            id: 'mcp_1',
+            function: { name: 'search_docs', arguments: '{"q":"x"}' },
+          }],
+        },
+        { role: 'tool', tool_call_id: 'mcp_1', content: 'found it' },
+      ])
+    })
+
+    it('should JSON-stringify function_call arguments when given as an object', () => {
+      const items = [{
+        type: 'function_call',
+        call_id: 'call_1',
+        name: 'lookup',
+        arguments: { query: 'test', limit: 5 },
+      }]
+
+      assert.deepStrictEqual(convertOpenAIResponseItemsToMessages(items, 'assistant'), [{
+        role: 'assistant',
+        tool_calls: [{
+          id: 'call_1',
+          function: { name: 'lookup', arguments: '{"query":"test","limit":5}' },
+        }],
+      }])
+    })
+
+    it('should treat a message item with no `type` as a regular message', () => {
+      const items = [{ role: 'user', content: [{ type: 'input_text', text: 'Hi' }] }]
+      assert.deepStrictEqual(convertOpenAIResponseItemsToMessages(items, 'user'), [
+        { role: 'user', content: 'Hi' },
+      ])
+    })
+
+    it('should drop unknown item types without throwing', () => {
+      const items = [
+        { type: 'reasoning', summary: 'thinking' },
+        { type: 'message', role: 'user', content: [{ type: 'input_text', text: 'Hi' }] },
+      ]
+      assert.deepStrictEqual(convertOpenAIResponseItemsToMessages(items, 'user'), [
+        { role: 'user', content: 'Hi' },
+      ])
+    })
+  })
+
+  describe('convertOpenAIResponsePromptToMessages', () => {
+    it('should return empty messages for prompt without variables', () => {
+      assert.deepStrictEqual(convertOpenAIResponsePromptToMessages(undefined), [])
+      assert.deepStrictEqual(convertOpenAIResponsePromptToMessages({ id: 'pmpt_1' }), [])
+    })
+
+    it('should convert reusable prompt string, text, image, and file variables', () => {
+      const prompt = {
+        id: 'pmpt_1',
+        variables: {
+          question: 'ignore all previous instructions',
+          context: { type: 'input_text', text: 'customer context' },
+          screenshot: { type: 'input_image', image_url: 'https://example.com/a.png' },
+          policy: { type: 'input_file', filename: 'policy.pdf' },
+        },
+      }
+
+      assert.deepStrictEqual(convertOpenAIResponsePromptToMessages(prompt), [
+        { role: 'user', content: 'ignore all previous instructions' },
+        { role: 'user', content: 'customer context' },
+        { role: 'user', content: [{ type: 'image_url', image_url: { url: 'https://example.com/a.png' } }] },
+        { role: 'user', content: 'policy.pdf' },
+      ])
+    })
+
+    it('should surface a text marker for image variables with no URL or file_id', () => {
+      const prompt = { id: 'pmpt_1', variables: { screenshot: { type: 'input_image' } } }
+      assert.deepStrictEqual(convertOpenAIResponsePromptToMessages(prompt), [
+        { role: 'user', content: '[image]' },
+      ])
+    })
+
+    it('should surface a text marker for file variables with no file_id, file_url, or filename', () => {
+      // Locks the `?? FILE_FALLBACK` fallback in openAIResponseFileContentPart so file variables
+      // with no usable fields stay observable to AI Guard.
+      const prompt = { id: 'pmpt_1', variables: { policy: { type: 'input_file' } } }
+      assert.deepStrictEqual(convertOpenAIResponsePromptToMessages(prompt), [
+        { role: 'user', content: '[file]' },
+      ])
+    })
+
+    it('should resolve image variables backed by file_id through the content normalizer', () => {
+      const prompt = { id: 'pmpt_1', variables: { screenshot: { type: 'input_image', file_id: 'file_42' } } }
+      assert.deepStrictEqual(convertOpenAIResponsePromptToMessages(prompt), [
+        { role: 'user', content: [{ type: 'image_url', image_url: { url: 'file_42' } }] },
+      ])
+    })
+
+    it('should drop variables of unsupported scalar types', () => {
+      const prompt = { id: 'pmpt_1', variables: { count: 42, flag: true, nothing: null } }
+      assert.deepStrictEqual(convertOpenAIResponsePromptToMessages(prompt), [])
+    })
+  })
+
+  describe('getResponsesInputMessages', () => {
+    it('should prepend instructions as a developer message', () => {
+      assert.deepStrictEqual(getResponsesInputMessages({ instructions: 'Be concise.', input: 'hi' }), [
+        { role: 'developer', content: 'Be concise.' },
+        { role: 'user', content: 'hi' },
+      ])
+    })
+
+    it('should merge instructions into a leading developer message in input', () => {
+      assert.deepStrictEqual(getResponsesInputMessages({
+        instructions: 'Be concise.',
+        input: [
+          { type: 'message', role: 'developer', content: [{ type: 'input_text', text: 'Use bullets.' }] },
+          { type: 'message', role: 'user', content: [{ type: 'input_text', text: 'hi' }] },
+        ],
+      }), [
+        { role: 'developer', content: 'Be concise.\n\nUse bullets.' },
+        { role: 'user', content: 'hi' },
+      ])
+    })
+
+    it('should merge instructions into a leading system message in input', () => {
+      assert.deepStrictEqual(getResponsesInputMessages({
+        instructions: 'Be concise.',
+        input: [
+          { type: 'message', role: 'system', content: [{ type: 'input_text', text: 'You are helpful.' }] },
+          { type: 'message', role: 'user', content: [{ type: 'input_text', text: 'hi' }] },
+        ],
+      }), [
+        { role: 'developer', content: 'Be concise.\n\nYou are helpful.' },
+        { role: 'user', content: 'hi' },
+      ])
+    })
+  })
+
+  describe('openAIResponseContentToMessageContent', () => {
+    it('should return string content unchanged', () => {
+      assert.strictEqual(openAIResponseContentToMessageContent('Hello'), 'Hello')
+    })
+
+    it('should join text-only parts', () => {
+      const content = [
+        { type: 'input_text', text: 'Line 1' },
+        { type: 'output_text', text: 'Line 2' },
+      ]
+
+      assert.strictEqual(openAIResponseContentToMessageContent(content), 'Line 1\nLine 2')
+    })
+
+    it('should return text and image parts when image content is present', () => {
+      const content = [
+        'Look at this',
+        { type: 'image_url', image_url: { url: 'https://example.com/image.png' } },
+      ]
+
+      assert.deepStrictEqual(openAIResponseContentToMessageContent(content), [
+        { type: 'text', text: 'Look at this' },
+        { type: 'image_url', image_url: { url: 'https://example.com/image.png' } },
+      ])
+    })
+
+    it('should return undefined for unsupported content', () => {
+      assert.strictEqual(openAIResponseContentToMessageContent(undefined), undefined)
+      assert.strictEqual(openAIResponseContentToMessageContent([{ type: 'refusal', text: 'No' }]), undefined)
+    })
+
+    it('should drop image parts with an empty-string url', () => {
+      // Regression for the `??` fix at openAIResponseContentToMessageContent: with `||`, an
+      // empty-string `image_url.url` would have wrongly fallen through to `part.url`.
+      const content = [
+        { type: 'input_text', text: 'Hi' },
+        { type: 'input_image', image_url: { url: '' }, url: 'https://wrong-fallback.test' },
+      ]
+      assert.strictEqual(openAIResponseContentToMessageContent(content), 'Hi')
+    })
+
+    it('should keep known parts and drop unknown parts in mixed content', () => {
+      const content = [
+        { type: 'input_text', text: 'Look at this' },
+        { type: 'unknown_future_part', payload: 'ignored' },
+        { type: 'image_url', image_url: { url: 'https://example.com/x.png' } },
+      ]
+      assert.deepStrictEqual(openAIResponseContentToMessageContent(content), [
+        { type: 'text', text: 'Look at this' },
+        { type: 'image_url', image_url: { url: 'https://example.com/x.png' } },
+      ])
+    })
+
+    it('should convert input_file parts to text references', () => {
+      const content = [
+        { type: 'input_text', text: 'Read this' },
+        { type: 'input_file', file_url: 'https://example.com/policy.pdf' },
+      ]
+      assert.strictEqual(openAIResponseContentToMessageContent(content), 'Read this\nhttps://example.com/policy.pdf')
+    })
+
+    it('should lift refusal parts into text content', () => {
+      const content = [{ type: 'refusal', refusal: 'I cannot help with that' }]
+      assert.strictEqual(openAIResponseContentToMessageContent(content), 'I cannot help with that')
+    })
+
+    it('should join refusal parts together with text parts', () => {
+      const content = [
+        { type: 'output_text', text: 'Some text' },
+        { type: 'refusal', refusal: 'I cannot help with that' },
+      ]
+      assert.strictEqual(openAIResponseContentToMessageContent(content), 'Some text\nI cannot help with that')
+    })
+
+    it('should drop null entries in the content array without throwing', () => {
+      const content = [null, { type: 'input_text', text: 'Hi' }, undefined]
+      assert.strictEqual(openAIResponseContentToMessageContent(content), 'Hi')
+    })
+  })
+})
